w mess

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: mcastrat <mcastrat@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 00:34:47 by mcastrat          #+#    #+#             */
-/*   Updated: 2024/10/21 02:55:01 by mcastrat         ###   ########.fr       */
+/*   Updated: 2024/10/21 09:28:02 by mcastrat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int			i;
+	size_t			i;
 	char		*cdest;
 	const char	*csrc;
 
@@ -31,7 +31,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n)
 	cdest[i] = '\0';
 	return (cdest);
 }
-/*
+/*va simplement copier un src dans une dest en fonction d une n taille et renvoyer la dest
 int main()
 {
 	char ye[10];
